refractor: browser storage functions -> browserStorageUtil.ts
