+++ frontend/src/util/helper.ts
@@ -118,77 +118,3 @@ export function mapBootcampToSaltData2(TrackName: string, bootcamp: BootcampResp
       displayName: "Fullstack " + TrackName 
   };
 }
-
-
-
-export async function openIndexedTemplatesDB(): Promise<IDBDatabase> {
-  return new Promise((resolve, reject) => {
-      const request = indexedDB.open('pdfCache', 1);
-
-      request.onupgradeneeded = (event) => {
-          const db = (event.target as IDBOpenDBRequest).result;
-          if (!db.objectStoreNames.contains('pdfs')) {
-              db.createObjectStore('pdfs', { keyPath: 'id' });
-          }
-      };
-
-      request.onsuccess = () => {
-          resolve(request.result);
-      };
-
-      request.onerror = () => {
-          reject(request.error);
-      };
-  });
-}
-
-export async function getFromIndexedTemplatesDB(db: IDBDatabase, key: string): Promise<any> {
-  return new Promise((resolve, reject) => {
-      const transaction = db.transaction('pdfs', 'readonly');
-      const store = transaction.objectStore('pdfs');
-      const request = store.get(key);
-
-      request.onsuccess = () => {
-          resolve(request.result);
-      };
-
-      request.onerror = () => {
-          reject(request.error);
-      };
-  });
-}
-
-export async function storeInIndexedTemplatesDB(db: IDBDatabase, key: string, data: any): Promise<void> {
-  return new Promise(async (resolve, reject) => {
-      const transaction = db.transaction('pdfs', 'readwrite');
-      const store = transaction.objectStore('pdfs');
-
-      const countRequest = store.count();
-      countRequest.onsuccess = async () => {
-          const count = countRequest.result;
-
-          if (count >= 25) {
-              const cursorRequest = store.openCursor();
-              cursorRequest.onsuccess = (event) => {
-                  const cursor = (event.target as IDBRequest<IDBCursorWithValue>).result;
-                  if (cursor) {
-                      store.delete(cursor.primaryKey);
-                      cursor.continue();
-                  }
-              };
-          }
-
-          const putRequest = store.put(data);
-          putRequest.onsuccess = () => {
-              resolve();
-          };
-          putRequest.onerror = () => {
-              reject(putRequest.error);
-          };
-      };
-
-      countRequest.onerror = () => {
-          reject(countRequest.error);
-      };
-  });
-}
